Remove tech radar origin from playbook
We're about to remove the Tech Radar because it's out of date and no
longer maintained. This change removes the origin, which will remove the
S3 bucket and and WAF configuration next time the terraform is run.

The variables here are used in the [playbook-distro branch of the
aws-hackney-common-terraform](https://github.com/LBHackney-IT/aws-hackney-common-terraform/tree/playbook-distro).
There it iterates over the contents of `origins`, so removing this
element won't break any references there.