Remove CODEOWNERS
In preparation for the global mandatory PR rules, this CODEOWNERS file
is out of date. We probably want a single team to own the repo, but
until we've figured that out we can remove all code owners. This means
the existing workflow of anyone in Hackney approving PRs can continue.