♻️  Refactor: use automatic sidebar generation
Previously to add a page you would need to create the file and then
create an entry in sidebars.js which references the ID. This is brittle,
and has resulting quite a few pages being created but not actually
included in the site. It also makes adding new content more difficult
than it needs to be, and requires a running site to curate the pages
with any confidence.

With this change the sidebar is generated automatically from the
directory structure within `docs`, so files can be moved around and
they'll automatically be picked up under the relevant heading. Where
a file had the same name as the section, renaming the file to `index.md`
causes the section to contain content, which may or may not be more
intuitive.

This change aimed to keep the new sidebar approximately the same as the
previous version, so introduces number ordering to file metadata
(`sidebar_position`) and
directories where ordering helps readability. It's been left off where
the order didn't seem like it had significance.

There are several files which were linked from within pages but didn't
exist in the sidebar. These have been placed with their linking
documents and now appear alongside them in the sidebar.

When moving the files we need to update a couple of things because we're
using relative paths:

- fix references to speech component
- fix image references